refactor: migrate SpaceID
