+++ packages/plugins/EVM/src/state/NameService/SpaceID.ts
@@ -1,34 +1,18 @@
 import type { NameServiceResolver } from '@masknet/web3-shared-base'
 import { NameServiceID } from '@masknet/shared-base'
-import SID, { getSidAddress } from '@siddomains/sidjs'
 import { ChainId } from '@masknet/web3-shared-evm'
-import { Web3StateSettings } from '../../settings/index.js'
+import { SpaceID } from '@masknet/web3-providers'
 
 export class SpaceID_Resolver implements NameServiceResolver {
-    private sid?: SID
-
-    private async init() {
-        if (this.sid) return
-
-        const web3 = await Web3StateSettings.value.Connection?.getWeb3?.({
-            chainId: ChainId.BSC,
-        })
-        this.sid = new SID({ provider: web3?.currentProvider, sidAddress: getSidAddress(ChainId.BSC) })
-    }
     public get id() {
         return NameServiceID.SpaceID
     }
 
     async lookup(name: string) {
-        if (!name.endsWith('.bnb')) return
-
-        await this.init()
-        return this.sid?.name(name).getAddress()
+        return SpaceID.lookup(ChainId.BSC, name)
     }
 
     async reverse(address: string) {
-        await this.init()
-        const result = await this.sid?.getName(address)
-        return result?.name
+        return SpaceID.reverse(ChainId.BSC, address)
     }
 }
