+++ packages/web3-providers/package.json
@@ -25,6 +25,7 @@
     "@masknet/web3-shared-solana": "workspace:^",
     "@metaplex-foundation/mpl-token-metadata": "^1.1.0",
     "@metaplex/js": "^4.11.7",
+    "@siddomains/sidjs": "^0.1.16",
     "@solana/buffer-layout": "^4.0.0",
     "@solana/spl-name-service": "^0.1.3",
     "@solana/spl-token": "^0.1.8",
