+++ packages/web3-providers/src/env.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="@masknet/global-types/webpack" />
+
 declare module 'ethjs-ens' {
     import { provider as Provider } from 'web3-core'
     export default class Ens {
@@ -11,4 +13,20 @@ declare module 'ethjs-ens' {
 
 declare module '@ensdomains/eth-ens-namehash' {
     export function hash(name: string): string
+}
+
+declare module '@siddomains/sidjs' {
+    import { Provider } from 'react'
+
+    export function getSidAddress(chain: number): number
+    export default class SID {
+        constructor(options: { provider: Provider; sidAddress: number }): void
+
+        name(name: string): Name
+        getName(address: string): Promise<{ name?: string }>
+    }
+
+    export class Name {
+        getAddress(): string
+    }
 }
